stationService with custom functions for stationModel
